+++ mimetic098/rfc822/group.cxx
@@ -33,9 +33,9 @@ Group::Group(const string& text)
 
 static string::size_type find_not_in_quote(const string& s, const string::value_type& c)
 {
-    int len = s.length();
+    size_t len = s.length();
     bool in_dquote = false;
-    for(int i =0; i < len; ++i)
+    for(int i = 0; i < len; ++i)
     {
         if(s[i] == '"')
             in_dquote = !in_dquote;
