+++ mimetic098/rfc822/messageid.cxx
@@ -19,7 +19,7 @@ MessageId::MessageId(uint32_t thread_id)
     if(!host.length())
         host = "unknown";
 
-    m_msgid = "m" + utils::int2str(time(0)) + "." + utils::int2str(getpid()) + 
+    m_msgid = "m" + utils::int2str((int)time(0)) + "." + utils::int2str(getpid()) + 
         "." + utils::int2str(thread_id) + 
         utils::int2str(++ms_sequence_number) + "@" + host;
 }
